Move handler e2e tests to handler package
We want the handler and operator tests to be in separate packages.
This paves the way for that.
