+++ test/e2e/handler/bonding_default_interface_test.go
@@ -1,4 +1,4 @@
-package e2e
+package handler
 
 import (
 	"fmt"
