+++ test/e2e/handler/conditions.go
@@ -1,4 +1,4 @@
-package e2e
+package handler
 
 import (
 	"context"
