+++ test/e2e/handler/default_bridged_network_test.go
@@ -1,4 +1,4 @@
-package e2e
+package handler
 
 import (
 	"context"
