+++ test/e2e/handler/multiple_policies_for_same_node_test.go
@@ -1,4 +1,4 @@
-package e2e
+package handler
 
 import (
 	"fmt"
