+++ test/e2e/handler/nnce_conditions_test.go
@@ -1,4 +1,4 @@
-package e2e
+package handler
 
 import (
 	"fmt"
