+++ test/e2e/handler/nns_update_timestamp_test.go
@@ -1,4 +1,4 @@
-package e2e
+package handler
 
 import (
 	"time"
