+++ test/e2e/handler/simple_bridge_and_bond_test.go
@@ -1,4 +1,4 @@
-package e2e
+package handler
 
 import (
 	"fmt"
