+++ test/e2e/handler/simple_ovs_bridge_test.go
@@ -1,4 +1,4 @@
-package e2e
+package handler
 
 import (
 	. "github.com/onsi/ginkgo"
