+++ test/e2e/handler/simple_vlan_and_ip_test.go
@@ -1,7 +1,8 @@
-package e2e
+package handler
 
 import (
 	"fmt"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
